Remove unwrap calls from the !resource command
Remove the `unwrap` calls from the `!resource` command so that the bot
cannot crash through a `panic!`. Utilise the `?` operator to propagate
the errors back up the chain and thus prevent the need for `unwrap`s.

Update the resource parsing to remove a leading space. Utilise borrows
to avoid allocating a `String` through `collect`, instead taking a slice
of the original string. This reduces memory consumption very slightly
and should be a little faster.

Remove a single space in the output message, such that there is no blank
line between the quote and the "... submitted the following resource",
reducing the space needed for each resource posted in the channel.