Fix a file system issue with !minutes
`!minutes` was failing to send the correct file for some reason,
potentially due to the file handle not being closed when `serenity`
received it and thus the data not being written (however flushing the
data didn't help either).

Uses a more basic approach where the file is deleted if it exists to
ensure a clean file is produced and the data is then written, with
`serenity` receiving the filename only.