Add a run.sh script for continuous delivery
Add a `run.sh` script to allow the bot to execute and update continuously. The
server's crontab contains `@reboot ~/pythia/run.sh`, which causes this script
to be executed after a reboot.

It then builds and runs `pythia` in the background, before sleeping for 30
minutes at a time. When it wakes up, it will automatically check whether
updates have been made on GitHub and kill the current instance and rebuild if
they have.

Paths to `cargo` are explicit, as the PATH variable is only updated with the
location of the binary after `~/.bashrc` has been read, so just using `cargo`
will fail as it cannot be found.