Upgrade the run.sh script with better logging
Upgrade the run.sh script to allow it to log better. Logs are stored in
`/root/.logs` and the activities of both run.sh and Pythia itself are logged
into files with timestamps.

Implement a shutdown command so that changes to run.sh will cause a server
restart and allow the new configuration to be used.